Fixed gui

+++ code/main/SceneAccountGUI.java
@@ -6,8 +6,6 @@
 package main;
 
 import javafx.application.Application;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.geometry.Insets;
@@ -17,22 +15,27 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
+import java.util.Optional;
+
 
 public class SceneAccountGUI extends Application {
-	User user = new User("User", "Password");
-    Stage theStage;
-    Scene scene;
-    SplitPane root;
-    ObservableList<Account> accountList = FXCollections.observableArrayList();; 
-    
+	private User user = new User("User", "Password");
+    private ObservableList<Account> accountList = FXCollections.observableArrayList();
+    private TextField userTextField;
+    private TextField appTextField;
+    private PasswordField pwBox;
+    private CheckBox autoPw;
+    private CheckBox selfPw;
+
     @Override
     public void start(Stage primaryStage) {
-        theStage = primaryStage;
+        Scene scene;
+        SplitPane root;
         root = new SplitPane();
         VBox left = new VBox();
         VBox right = new VBox();
-        CheckBox autoPw = new CheckBox();
-        CheckBox selfPw = new CheckBox();
+        autoPw = new CheckBox();
+        selfPw = new CheckBox();
         root.getItems().addAll(left, right);
         root.setDividerPosition(0, 0.45);
 
@@ -59,35 +62,33 @@ public class SceneAccountGUI extends Application {
         modifyBox.getChildren().addAll(modifyButton, modifyErr);
 
         Label userName = new Label("User Name:");
-        TextField userTextField = new TextField();
+        userTextField = new TextField();
 
         autoPw.setText("Automatically generate password");
         selfPw.setText("Set password yourself");
 
         Label password = new Label("Password: ");
-        PasswordField pwBox = new PasswordField();
+        pwBox = new PasswordField();
         pwBox.setEditable(false);
 
         Label appName = new Label("Application:");
-        TextField appTextField = new TextField();
+        appTextField = new TextField();
 
         VBox vbox = new VBox();
         vbox.setSpacing(30);
 
         vbox.getChildren().addAll(userName, userTextField, password, autoPw, selfPw, pwBox, appName, appTextField, addBox, deleteBox, modifyBox);
 
-
         left.getChildren().addAll(vbox);
 
-        TableView<Account> tvAccount = new TableView<Account>(accountList);
-
+        TableView<Account> tvAccount = new TableView<>(accountList);
+        ScrollBar scrollBar = new ScrollBar();
         TableColumn<Account, String> uName = new TableColumn<>("User Name");
-        uName.setCellValueFactory(new PropertyValueFactory<>("Username"));
-
         TableColumn<Account, String> pw = new TableColumn<>("Password");
-        pw.setCellValueFactory(new PropertyValueFactory<>("password"));
-
         TableColumn<Account, String> aName = new TableColumn<>("Application name");
+
+        uName.setCellValueFactory(new PropertyValueFactory<>("Username"));
+        pw.setCellValueFactory(new PropertyValueFactory<>("password"));
         aName.setCellValueFactory(new PropertyValueFactory<>("Appname"));
 
         tvAccount.getColumns().addAll(uName, pw, aName);
@@ -124,21 +125,17 @@ public class SceneAccountGUI extends Application {
         });
 
         addButton.setOnAction(event -> {
-            Account acc = new Account ("","", "");
-            if (autoPw.isSelected())
-                acc = new Account(userTextField.getText().trim(), 20, appTextField.getText().trim());
-            else if (selfPw.isSelected())
-                acc = new Account(userTextField.getText().trim(), pwBox.getText().trim(), appTextField.getText().trim());
+            Account acc = pwOption(autoPw, selfPw);
             if (user.addAccount(acc)) {
                 accountList.add(acc);
                 addDuplicate.setText("");
             }
             else
-                addDuplicate.setText("  You cannot add a duplicated or blank account");
+                addDuplicate.setText("You cannot add a duplicated or blank account");
             userTextField.setText("");
             pwBox.setText("");
             appTextField.setText("");
-            textFieldsToVoid(userTextField,pwBox,appTextField);
+            inputToVoid(userTextField,pwBox,appTextField);
             deleteErr.setText("");
             modifyErr.setText("");
         });
@@ -146,37 +143,38 @@ public class SceneAccountGUI extends Application {
         deleteButton.setOnAction(event -> {
             Account deletedItem = tvAccount.getSelectionModel().getSelectedItem();
             if (deletedItem != null) {
-                user.deleteAccount(deletedItem.getId());
-                tvAccount.getItems().remove(deletedItem);
-                deleteErr.setText("");
-            }
-            else
+                if (alertMessage("delete")) {
+                    user.deleteAccount(deletedItem.getId());
+                    tvAccount.getItems().remove(deletedItem);
+                    deleteErr.setText("");
+                }
+            } else
                 deleteErr.setText("Please choose an account you want to delete.");
             tvAccount.getSelectionModel().clearSelection();
-            textFieldsToVoid(userTextField,pwBox,appTextField);
+            inputToVoid(userTextField, pwBox, appTextField);
             modifyErr.setText("");
             addDuplicate.setText("");
         });
 
         modifyButton.setOnAction(event -> {
             Account modifiedItem  = tvAccount.getSelectionModel().getSelectedItem();
-            Account newItem = new Account("","","");
             if (modifiedItem != null) {
-                if (autoPw.isSelected())
-                    newItem = new Account(userTextField.getText().trim(), 20, appTextField.getText().trim());
-                else if (selfPw.isSelected())
-                    newItem = new Account(userTextField.getText().trim(), pwBox.getText().trim(), appTextField.getText().trim());
+                Account newItem = pwOption(autoPw, selfPw);
                 if (user.modifyAccount(modifiedItem.getId(), newItem)) {
-                    tvAccount.getItems().set(accountList.indexOf(modifiedItem), newItem);
-                    modifyErr.setText("");
+                    if (alertMessage("modify")) {
+                        tvAccount.getItems().set(accountList.indexOf(modifiedItem), newItem);
+                        modifyErr.setText("");
+                    }
                 }
-                else
-                    modifyErr.setText("  You cannot modify this entry with a duplicated or blank account");
+                else {
+                    modifyErr.setText("You cannot modify this entry with a duplicated or blank account. If you omit the password options, the account is considered blank.");
+                    modifyErr.setWrapText(true);
+                }
             }
             else
                 modifyErr.setText("Please choose an account you want to modify.");
             tvAccount.getSelectionModel().clearSelection();
-            textFieldsToVoid(userTextField,pwBox,appTextField);
+            inputToVoid(userTextField,pwBox,appTextField);
             addDuplicate.setText("");
             deleteErr.setText("");
         });
@@ -189,10 +187,32 @@ public class SceneAccountGUI extends Application {
         });
     }
 
-    public void textFieldsToVoid(TextField userName, TextField pw, TextField appName) {
+    private void inputToVoid(TextField userName, TextField pw, TextField appName) {
         userName.setText("");
         pw.setText("");
         appName.setText("");
+        autoPw.setSelected(false);
+        selfPw.setSelected(false);
+    }
+
+    private Account pwOption (CheckBox autoPw, CheckBox selfPw) {
+        Account newItem;
+        if (autoPw.isSelected())
+            newItem = new Account(userTextField.getText().trim(), 20, appTextField.getText().trim());
+        else if (selfPw.isSelected())
+            newItem = new Account(userTextField.getText().trim(), pwBox.getText().trim(), appTextField.getText().trim());
+        else
+            newItem = new Account ("","","");
+        return newItem;
+    }
+
+    private boolean alertMessage(String message) {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation");
+        alert.setHeaderText(null);
+        alert.setContentText("Are you sure you want to permanently " + message + " this account?");
+        Optional<ButtonType> action = alert.showAndWait();
+        return action.get() == ButtonType.OK;
     }
 
    /**
